requirements.txt
autopep8==1.6.0
django-bootstrap-form==3.4
django-extensions==3.1.5
django-filter==22.1
Django==4.0.5
djhtml==1.5.1
isort==5.10.1
Pillow==9.1.1
python-decouple==3.6
django-localflavor==3.1
